+++ bin/di/container.ts
@@ -235,7 +235,6 @@ export const DIServiceKeys = {
     VERATOWN: "veratown",
     KEYPAD_DEFINITION_SERVICE: "keypadDefinitionService",
     KEYPAD_ACCESS_SERVICE: "keypadAccessService",
-    KEYPAD_ACCESS_GROUP_MANAGER: "keypadAccessGroupManager",
     KEYPAD_DOOR_SYSTEM: "keypadDoorSystem",
     KIDNAPPERS_GAME_PERSISTENCE: "kidnappersGamePersistence",
     KIDNAPPERS_GAME_LIFECYCLE_SERVICE: "kidnappersGameLifecycleService",